pyproject.toml
[project]
name = "genesis-agi"
version = "0.1.0"
description = "BabyAGIライクな自己改善型タスク管理システム"
authors = [{ name = "Your Name" }]
dependencies = [
    "openai>=1.0.0",
    "redis>=5.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.0.0",
    "numpy>=2.0.2",
    "msgpack>=1.1.0",
]
requires-python = ">=3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["genesis_agi"]

[tool.ruff]
line-length = 88
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]
ignore = []

[tool.ruff.isort]
known-first-party = ["genesis_agi"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=genesis_agi"
